refine && merge
