+++ integration_tests/src/main/python/collection_ops_test.py
@@ -30,17 +30,27 @@ non_nested_array_gens = [ArrayGen(sub_gen, nullable=nullable)
 @pytest.mark.parametrize('data_gen', non_nested_array_gens, ids=idfn)
 def test_concat_list(data_gen):
     assert_gpu_and_cpu_are_equal_collect(
+        lambda spark: binary_op_df(spark, data_gen).selectExpr('concat(a)'))
+
+    assert_gpu_and_cpu_are_equal_collect(
         lambda spark: binary_op_df(spark, data_gen).selectExpr('concat(a, b)'))
 
     assert_gpu_and_cpu_are_equal_collect(
         lambda spark: three_col_df(spark, data_gen, data_gen, data_gen
                                    ).selectExpr('concat(a, b, c)'))
 
+def test_empty_concat_list():
+    assert_gpu_and_cpu_are_equal_collect(
+        lambda spark: binary_op_df(spark, ArrayGen(LongGen())).selectExpr('concat()'))
+
 def test_concat_string():
     gen = mk_str_gen('.{0,5}')
     (s1, s2) = gen_scalars(gen, 2, force_no_nulls=True)
     assert_gpu_and_cpu_are_equal_collect(
             lambda spark: binary_op_df(spark, gen).select(
+                f.concat(),
+                f.concat(f.col('a')),
+                f.concat(s1),
                 f.concat(f.col('a'), f.col('b')),
                 f.concat(f.col('a'), f.col('b'), f.col('a')),
                 f.concat(s1, f.col('b')),
